Make Objects automatically an ObjectList of more than one object of the same name is created.
This removes the necessity to create the APP $drawing call in the code. We therefore removed the §drawing array from the code scope completely and capture the objects in the interpreters callback. All in all this simplifies the code and prevents index shenanigans when generating and manipulating code, because the hidden statements manipulate how the code indexes are assigned to the visible statements.