Prevent nested loops.
The runtime supports this. But right now, when iterating over a list of values using FOREACH we're using the list register name as the single value register within the loop. This means we're masking the list register and cannot iterate over it within the same loop. We could use a iteration variable by doing FOREACH v, list. But we don't to make it easier for the user.