Make constructed Object available in the global scope.
Object created in a forEach-loop are by default not global as forEach opens a new stackFrame. Using the objectCallback of the GfxInterpreter we make the objects available in the global scope. This allows the user to use these objects, most likely the object-list after the loop is done, e.g. for changing style. As always object list will delegate everything to the last object in its list. This behavior can be toggled on/off using the AppConfig.Runtime.grObjectsGlobal option. It is turned on right now.