Move Library to the right sidebar.
This allows for more layout options and we can keep it always visible, which is a good thing. Inserting a drawing is a tool like all other tools and should be available all the time.

Other ideas: Map inster drawing to a key and toolbar button and popup the library when activated...