Switch ForEach to an explicit iterator register.
We used the name of the list register before. That made it easy for the user to comprehend. But it created loads of problems, especially in nested forEach-loops. We now automatically create a new iterator for each forEach. When using Update-Statement, we automatically recognize, if we're inside a forEach and use the iterator automatically.